lint fix

+++ src/collections/executor.c
@@ -40,7 +40,7 @@ _z_executor_spin_result_t _z_executor_get_next_fut(_z_executor_t *executor, _z_f
         uint64_t wake_up_time_ms = _z_fut_schedule_get_wake_up_time_ms(
             _z_fut_data_hmap_node_at(&executor->_tasks, *sleeping_idx_ptr)->val._schedule);
         z_clock_t wake_up_time = executor->_epoch;
-        z_clock_advance_ms(&wake_up_time, wake_up_time_ms);
+        z_clock_advance_ms(&wake_up_time, (unsigned long)wake_up_time_ms);
         if (zp_clock_elapsed_ms_since(&now, &wake_up_time) > 0) {
             // The sleeping task is ready to execute
             _z_fut_data_hmap_index_t sleeping_idx;
